Task-19-FedoraHappiness/requirements/base.txt
#django-haystack doesn't have support for django2.1 yet
Django==2.0

# Basic Django extensions
django-extensions==1.6.1
django-braces==1.8.1
django-crispy-forms==1.7.2
django-model-utils==2.4
dogslow==1.2                       # For logging details about very slow requests
six>=1.11.0                          # Dependency of django-extensions

# Basic Python libraries
Unipath==1.1          # Better path handling
pytz==2015.7          # Timezone data
ipython==7.3.0        # Prettier python prompt

# For generating docs - sphinx and dependencies
Sphinx==1.3.6
Jinja2==2.8
MarkupSafe==0.23
Pygments==2.1.2
docutils==0.12

# PEP8 validation with dependencies
flake8==2.5.4
mccabe==0.4.0
pyflakes==1.0.0
pep8==1.7.0

# Misc
urllib3==1.24.2       # Dependency for requests 
python-dateutil==2.5.0
factory-boy==2.9.2
mozilla-django-oidc==1.2.2
fedora-messaging>=1.4.0
happinesspacket-schema>=1.0.0
celery[redis]==4.2.1
django-celery-email==2.0.1
psycopg2==2.7.5 # PostgreSQL driver
bleach>=3.1.0

# Search engine
Whoosh==2.7.4
django-haystack==2.8.1

#python-fedora for f-a-s API
python-fedora==0.10.0

#Dependency for normalizing emails
email-normalize==0.2.1

# Dependency for YAML file
pyyaml==5.1

# WYSIWYG addition
django-ckeditor==5.7.0

#Testing using pytest
pytest-django==3.5.0
pytest==4.6.3
pytest-cov==2.7.1
pytest-mock==1.10.4